+++ CHANGELOG.md
@@ -108,6 +108,7 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 
 * Records load development task
 * GitLab issues from additional information tab (now in admin tab)
+* Administration metadata as a record requirement
 
 ## [0.4.0] - 2025-10-24